requirements.txt
streamlit
pypdf2
langchain
langchain-community
python-dotenv
faiss-cpu
huggingface_hub
InstructorEmbedding
sentence-transformers==2.2.2
langchain-google-genai